Disable migrations for tests
This dramatically speeds up tests. The downside is that data
migrations aren't run for tests, which is just something we have
be aware of.

So, for example, we can't set up fixtures using data migrations.
(Ironically I did that in the first PR I made, but it doesn't
seem to break the tests so ¯\_(ツ)_/¯)