Fix emails not sending with custom template
This was introduced in the Django 1.11 upgrade. Django 1.11 requires
context for template rendering to be a dict rather than a Context()
object, so this was changed as part of upgrade.

Weirdly though, this doesn't apply when using Template() to render
a string. So, I've refactored so Context() is used for custom
templates, but not when rendering default template off disk.