Attempt to fix flakey email tests
Sometimes these tests don't send out emails, e.g. in #540.

This is a guess at a fix -- the test and the real implementation
use a different definition of "today" so I'm guessing it fails
intermittently when the timezone and UTC version of today differs.

These opens up a can of timezone worms -- various tasks use
different definitions of today, most importantly the billing one.

I haven't thought about this in much detail but I think this will
fix the test at least. 😬