Fix npm packages in Docker dev environment
This is a gnarly little problem with no ideal solution. We mount
the node_modules directory as a volume so that the version from
the image is used. But, that means that if the image is updated,
the volume is remounted and the version of the packages from
the previous image is used (I think??). Anyway -- whatever
the precise issue, doing `npm install` before running in dev fixes
the issue.